requirement.txt
opencv-python-headless
mysql-connector-python
Pillow
numpy
